requirements.txt
Django==1.11.2
Pillow==4.1.1
certifi==2017.4.17
chardet==3.0.4
django-embed-video==1.1.2
idna==2.5
olefile==0.44
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
